Schematic updates
Removed rc lowpass filter, since there's plenty of filtering in
software, and the RSSI signals from the RX5808 tend to fluctuate wildly
anyway. This hardware filtering was overkill, and would have made the
build tougher, more components to place, etc.

Add 1M trimmer pots for each receiver, which will allow tuning to match
the RSSI levels between the receivers. I added this because I noticed
while testing with all receivers tuned to the same channel, that the
peaks from some receivers were much higher (or lower) than that of
others. The pot is a somewhat crude approach, and it's not a true
voltage divider, but it should accomplish what we need with a minimum
part count. The pots can also be left off the board if desired, to get
the full voltage from the RSSI signal.

Add 3.3V voltage regulator to power to RX5808 modules, so that they will
run a bit cooler.

Remove flow control signals from the UART header. This avoids build
complications, and at 19200 baud, flow control shouldn't be necessary
anyway.

Remove buzzer.

Route spare Teensy signals to a 6 pin header.

Replace the Teensy 3.2 symbol with one that better matches the layout.
The previous Teensy symbol was a source of confusion and headaches when
working on the breadboard.