Auto-detect installed modules
Modules are now auto-detected upon initialization. If a module is not
installed, then its frequency can't be changed, and it can't be enabled
via the serial protocol.

The detection method involves enabling an internal pull-down resistor on
each module's RSSI signal, and sensing the voltage. When a module is
installed, the signal will remain above a certain threshold because the
module is driving the signal at or above 500mv. If no module is
connected, then the signal will read near 0v.